# controlplane/gen/.openapi-3.0.yaml
openapi: 3.0.3
info:
  title: LangSmith Deployment
  version: 0.1.0
  description: "LangGraph Cloud-compatible API surface implemented by duragraph.\n\
    \nNote: paths are documented as bare (`/assistants`, `/threads`, `/runs`,\netc.)\
    \ to mirror the upstream LangGraph spec. The duragraph engine\nserves these paths\
    \ under the `/api/v1` base \u2014 declared via the\n`servers:` block below. Clients\
    \ should construct request URLs as\n`<server.url><path>`, e.g. `http://localhost:8081/api/v1/runs`.\n"
servers:
- url: http://localhost:8081/api/v1
  description: Local development server
- url: https://api.duragraph.ai/api/v1
  description: "Production server (placeholder \u2014 replace with deployment host)"
...
paths:
...
  /threads/{thread_id}/stream:
    get:
      tags:
      - Threads
      summary: Join Thread Stream
      description: This endpoint streams output in real-time from a thread. The stream
        will include the output of each run executed sequentially on the thread and
        will remain open indefinitely. It is the responsibility of the calling client
        to close the connection.
      operationId: join_thread_stream_threads__thread_id__stream_get
      parameters:
      - description: The ID of the thread.
        required: true
        schema:
          type: string
          format: uuid
          title: Thread Id
          description: The ID of the thread.
        name: thread_id
        in: path
      - required: false
        schema:
          type: string
          title: Last Event ID
          description: The ID of the last event received. Used to resume streaming
            from a specific point. Pass '-' to resume from the beginning.
        name: Last-Event-ID
        in: header
      - required: false
        schema:
          default:
          - run_modes
          title: Stream Modes
          description: Stream modes to control which events are returned. 'lifecycle'
            returns only run start/end events, 'run_modes' returns all run events
            (default behavior), 'state_update' returns only state update events.
        name: stream_modes
        in: query
      responses:
        '200':
          description: Success
          content:
            text/event-stream:
              schema:
                type: string
                description: 'The server will send a stream of events in SSE format.


                  **Example event**:


                  id: 1


                  event: message


                  data: {}'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /threads/{thread_id}/runs:
    get:
      tags:
      - Thread Runs
      summary: List Runs
      description: List runs for a thread.
      operationId: list_runs_http_threads__thread_id__runs_get
      parameters:
      - description: The ID of the thread.
        required: true
        schema:
          type: string
          format: uuid
          title: Thread Id
          description: The ID of the thread.
        name: thread_id
        in: path
      - required: false
        schema:
          type: integer
          title: Limit
          default: 10
        name: limit
        in: query
      - required: false
        schema:
          type: integer
          title: Offset
          default: 0
        name: offset
        in: query
      - required: false
        schema:
          type: string
          enum:
          - pending
          - error
          - success
          - timeout
          - interrupted
        name: status
        in: query
      - required: false
        schema:
          type: array
          items:
            type: string
            enum:
            - run_id
            - thread_id
            - assistant_id
            - created_at
            - updated_at
            - status
            - metadata
            - kwargs
            - multitask_strategy
          title: Select
          description: Specify which fields to return. If not provided, all fields
            are returned.
        name: select
        in: query
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Run'
                type: array
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
      - Thread Runs
      summary: Create Background Run
      description: Create a run in existing thread, return the run ID immediately.
        Don't wait for the final run output.
      operationId: create_run_threads__thread_id__runs_post
      parameters:
      - description: The ID of the thread.
        required: true
        schema:
          type: string
          format: uuid
          title: Thread Id
          description: The ID of the thread.
        name: thread_id
        in: path
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunCreateStateful'
        required: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Run'
          headers:
            Content-Location:
              description: The URL of the run that was created. Can be used to later
                join the stream.
              schema:
                type: string
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /threads/{thread_id}/runs/stream:
    post:
      tags:
      - Thread Runs
      summary: Create Run, Stream Output
      description: Create a run in existing thread. Stream the output.
      operationId: stream_run_threads__thread_id__runs_stream_post
      parameters:
      - description: The ID of the thread.
        required: true
        schema:
          type: string
          format: uuid
          title: Thread Id
          description: The ID of the thread.
        name: thread_id
        in: path
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunCreateStateful'
        required: true
      responses:
        '200':
          description: Success
          content:
            text/event-stream:
              schema:
                type: string
                description: 'The server will send a stream of events in SSE format.


                  **Example event**:


                  id: 1


                  event: message


                  data: {}'
          headers:
            Content-Location:
              description: The URL of the run that was created. Can be used to later
                join the stream.
              schema:
                type: string
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /threads/{thread_id}/runs/{run_id}:
    get:
      tags:
      - Thread Runs
      summary: Get Run
      description: Get a run by ID.
      operationId: get_run_http_threads__thread_id__runs__run_id__get
      parameters:
      - description: The ID of the thread.
        required: true
        schema:
          type: string
          format: uuid
          title: Thread Id
          description: The ID of the thread.
        name: thread_id
        in: path
      - description: The ID of the run.
        required: true
        schema:
          type: string
          format: uuid
          title: Run Id
          description: The ID of the run.
        name: run_id
        in: path
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Run'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
      - Thread Runs
      summary: Delete Run
      description: Delete a run by ID.
      operationId: delete_run_threads__thread_id__runs__run_id__delete
      parameters:
      - description: The ID of the thread.
        required: true
        schema:
          type: string
          format: uuid
          title: Thread Id
          description: The ID of the thread.
        name: thread_id
        in: path
      - description: The ID of the run.
        required: true
        schema:
          type: string
          format: uuid
          title: Run Id
          description: The ID of the run.
        name: run_id
        in: path
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema: {}
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /threads/{thread_id}/runs/{run_id}/join:
    get:
      tags:
      - Thread Runs
      summary: Join Run
      description: Wait for a run to finish.
      operationId: join_run_http_threads__thread_id__runs__run_id__join_get
      parameters:
      - description: The ID of the thread.
        required: true
        schema:
          type: string
          format: uuid
          title: Thread Id
          description: The ID of the thread.
        name: thread_id
        in: path
      - description: The ID of the run.
        required: true
        schema:
          type: string
          format: uuid
          title: Run Id
          description: The ID of the run.
        name: run_id
        in: path
      - required: false
        schema:
          type: boolean
          title: Cancel on Disconnect
          description: If true, the run will be cancelled if the client disconnects.
          default: false
        name: cancel_on_disconnect
        in: query
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema: {}
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /threads/{thread_id}/runs/{run_id}/stream:
    get:
      tags:
      - Thread Runs
      summary: Join Run Stream
      description: Join a run stream. This endpoint streams output in real-time from
        a run similar to the /threads/__THREAD_ID__/runs/stream endpoint. If the run
        has been created with `stream_resumable=true`, the stream can be resumed from
        the last seen event ID.
      operationId: stream_run_http_threads__thread_id__runs__run_id__join_get
      parameters:
      - description: The ID of the thread.
        required: true
        schema:
          type: string
          format: uuid
          title: Thread Id
          description: The ID of the thread.
        name: thread_id
        in: path
      - description: The ID of the run.
        required: true
        schema:
          type: string
          format: uuid
          title: Run Id
          description: The ID of the run.
        name: run_id
        in: path
      - required: false
        schema:
          type: string
          title: Last Event ID
          description: The ID of the last event received. Set to -1 if you want to
            stream all events. Requires `stream_resumable=true` to be set when creating
            the run.
        name: Last-Event-ID
        in: header
      - required: false
        schema:
          type: string
          title: Stream Mode
          description: The mode to stream the run in. If not provided, the default
            mode will be used.
        name: stream_mode
        in: query
      - required: false
        schema:
          type: boolean
          title: Cancel On Disconnect
          description: If true, the run will be cancelled if the client disconnects.
          default: false
        name: cancel_on_disconnect
        in: query
      responses:
        '200':
          description: Success
          content:
            text/event-stream:
              schema:
                type: string
                description: 'The server will send a stream of events in SSE format.


                  **Example event**:


                  id: 1


                  event: message


                  data: {}'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /threads/{thread_id}/runs/{run_id}/cancel:
    post:
      tags:
      - Thread Runs
      summary: Cancel Run
      operationId: cancel_run_http_threads__thread_id__runs__run_id__cancel_post
      parameters:
      - description: The ID of the thread.
        required: true
        schema:
          type: string
          format: uuid
          title: Thread Id
          description: The ID of the thread.
        name: thread_id
        in: path
      - description: The ID of the run.
        required: true
        schema:
          type: string
          format: uuid
          title: Run Id
          description: The ID of the run.
        name: run_id
        in: path
      - required: false
        schema:
          type: boolean
          title: Wait
          default: false
        name: wait
        in: query
      - description: Action to take when cancelling the run. Possible values are `interrupt`
          or `rollback`. `interrupt` will simply cancel the run. `rollback` will cancel
          the run and delete the run and associated checkpoints afterwards.
        required: false
        schema:
          type: string
          enum:
          - interrupt
          - rollback
          title: Action
          default: interrupt
        name: action
        in: query
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema: {}
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /runs/stream:
    post:
      tags:
      - Stateless Runs
      summary: Create Run, Stream Output
      description: Create a run in a new thread, stream the output.
      operationId: stream_run_stateless_runs_stream_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunCreateStateless'
        required: true
      responses:
        '200':
          description: Success
          content:
            text/event-stream:
              schema:
                type: string
                description: 'The server will send a stream of events in SSE format.


                  **Example event**:


                  id: 1


                  event: message


                  data: {}'
          headers:
            Content-Location:
              description: The URL of the run that was created. Can be used to later
                join the stream.
              schema:
                type: string
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /runs/cancel:
    post:
      tags:
      - Thread Runs
      summary: Cancel Runs
      description: Cancel one or more runs. Can cancel runs by thread ID and run IDs,
        or by status filter.
      operationId: cancel_runs_post
      parameters:
      - description: Action to take when cancelling the run. Possible values are `interrupt`
          or `rollback`. `interrupt` will simply cancel the run. `rollback` will cancel
          the run and delete the run and associated checkpoints afterwards.
        required: false
        schema:
          type: string
          enum:
          - interrupt
          - rollback
          title: Action
          default: interrupt
        name: action
        in: query
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunsCancel'
        required: true
      responses:
        '204':
          description: Success - Runs cancelled
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /runs/wait:
    post:
      tags:
      - Stateless Runs
      summary: Create Run, Wait for Output
      description: Create a run in a new thread. Wait for the final output and then
        return it.
      operationId: wait_run_stateless_runs_wait_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunCreateStateless'
        required: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema: {}
          headers:
            Content-Location:
              description: The URL of the run that was created. Can be used to later
                join the stream.
              schema:
                type: string
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    RunsCancel:
      type: object
      title: RunsCancel
      description: Payload for cancelling runs.
      properties:
        status:
          type: string
          enum:
          - pending
          - running
          - all
          title: Status
          description: Filter runs by status to cancel. Must be one of 'pending',
            'running', or 'all'.
        thread_id:
          type: string
          format: uuid
          title: Thread Id
          description: The ID of the thread containing runs to cancel.
        run_ids:
          type: array
          items:
            type: string
            format: uuid
          title: Run Ids
          description: List of run IDs to cancel.
...
    ErrorResponse:
      type: string
      title: ErrorResponse
      description: Error message returned from the server
...